Add docs and split modbus handling
